dashboards/requirements.txt
alabaster==0.7.13
altair==4.2.0
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.10.1
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Automat==22.10.0
autopep8==2.0.1
awscli==1.27.50
azure-core==1.26.2
azure-identity==1.12.0
azure-kusto-data==4.0.2
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.10.0
bleach==3.3.0
blinker==1.5
boto3==1.16.7
botocore==1.29.50
branca==0.4.2
bs4==0.0.1
cachetools==5.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.4
commonmark==0.9.1
configparser==5.3.0
constantly==15.1.0
country-converter==0.8.0
coverage==7.0.5
cryptography==39.0.0
cssselect==1.2.0
cycler==0.10.0
Cython==0.29.23
dask==2023.3.0
DateTime==4.3
dbus-python==1.2.16
decorator==5.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.4.0
docutils==0.16
entrypoints==0.3
exceptiongroup==1.1.0
facets-overview==1.0.0
fake-useragent==1.1.1
filelock==3.0.12
findspark==2.0.0
Fiona==1.8.20
flake8==6.0.0
Flask==2.2.2
flexpolyline==0.1.0
folium==0.14.0
fsspec==2023.3.0
geopandas==0.10.2
gitdb==4.0.9
GitPython==3.1.29
google-services-api==1.5.0
googlemaps==4.5.3
h11==0.14.0
httmock==1.4.0
http-request-randomizer==1.3.2
hyperlink==21.0.0
idna==2.10
ijson==3.2.0.post0
imagesize==1.4.1
importlib-metadata==5.0.0
importlib-resources==5.10.2
incremental==22.10.0
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.13.3
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
koalas==1.8.1
locket==1.0.0
lxml==4.9.2
Mako==1.1.0
maps-analytics-utils==1.1
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.4.2
mccabe==0.7.0
mistune==0.8.4
msal==1.20.0
msal-extensions==1.0.0
munch==2.5.0
nbclassic==0.3.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.2
outcome==1.2.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.12.0
parsel==1.7.0
parso==0.7.0
partd==1.3.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==5.1.0
polyline==1.4.0
portalocker==2.6.0
prometheus-client==0.10.1
prompt-toolkit==3.0.17
Protego==0.2.1
protobuf==3.17.2
psutil==5.9.4
psycopg2-binary==2.8.5
ptyprocess==0.7.0
py==1.11.0
py4j==0.10.9
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.20
pydeck==0.8.0
PyDispatcher==2.0.6
pyflakes==3.0.1
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==2.6.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==2.4.7
pyproj==3.3.0
pyrsistent==0.17.3
pyspark==3.1.2
python-dateutil==2.8.1
python-dotenv==0.21.0
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
pyzmq==20.0.0
queuelib==1.6.2
requests==2.25.1
requests-file==1.5.1
requests-unixsocket==0.2.0
retry==0.9.2
rich==12.6.0
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==0.24.1
scipy==1.6.2
Scrapy==2.7.1
seaborn==0.11.1
selenium==4.7.2
semver==2.13.0
Send2Trash==1.5.0
service-identity==21.1.0
Shapely==1.8.0
six==1.15.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sphinx==6.1.3
sphinxcontrib-applehelp==1.0.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.10
statsmodels==0.12.2
streamlit==1.14.1
streamlit-folium==0.10.0
tenacity==8.0.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.0.5
trimesh==3.18.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
typing==3.7.4.3
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.25.11
validators==0.20.0
virtualenv==20.4.1
w3lib==2.1.1
watchdog==2.1.9
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
widgetsnbextension==3.5.1
wsproto==1.2.0
zipp==3.10.0
zope.interface==5.4.0
